_archive/billing-doc-analyzer/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "billing-doc-analyzer"
version = "0.1.0"
description = "AI agent that analyzes billing documents (bills, letters, emails) from service providers"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "pdfplumber>=0.10.0",
    "pydantic>=2.0.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.18.0"]
dev = ["pytest>=8.0.0", "httpx>=0.26.0"]

[tool.setuptools.packages.find]
where = ["src"]
